Rename repository
Unforunately, It came to my knowledge that this repository
promoted a phenomenon where DevOps interviews became a trivia game
where people think it's normal to throw 20 random short questions
like "what is fork()" or "which tools would you use for each
of the following areas?" and this was not my intention.

To explcitly state this repository doesn't represents real DevOps
interview questions I've decided to rename it.